added transform trigger endpoint and list endpoints, added query repository for outputdata
