+++ backend/infrastructures/apps/load/adapters/__init__.py
@@ -1,6 +1,8 @@
 from .units_of_work import DjangoDataUnitOfWork
 from .repositories import DjangoDataDomainRepository
+from .repositories import DjangoDataQueryRepository
 __all__ = [
     "DjangoDataUnitOfWork",
     "DjangoDataDomainRepository",
+    "DjangoDataQueryRepository",
 ]
