+++ backend/infrastructures/apps/load/adapters/queries.py
@@ -4,18 +4,18 @@ from typing import Any
 
 from django.contrib.auth import get_user_model
 
-from modules.load.domain.ports import repositories as domain_repositories 
+from modules.load.services.queries import ports as query_ports 
 from modules.data.domain import value_objects as data_value_objects
 
 
 logger = logging.getLogger(__name__)
 
-class DjangoDataDomainRepository(domain_repositories.AbstractDataDomainRepository):
+class DjangoDataQueryRepository(query_ports.AbstractDataQueryRepository):
     """
     See description of parent class to get more details.
     """
-    def create(self, data: list[data_value_objects.OutputData]) -> None:
+    def list(self) -> list[data_value_objects.OutputData]:
         #TODO 4: investigate how to connect to postgres database 
-        #TODO this is supposed to raise DataCreationException
-        logger.info("database load action here")
+        #TODO this is supposed to raise DataAccessException
+        logger.info("database list action here")
