+++ backend/infrastructures/config/settings/injections_configurations.py
@@ -2,7 +2,8 @@ import inject
 from django.conf import settings
 
 
-from ...apps.load.adapters import repositories as load_repositories
+from ...apps.load.adapters import repositories as domain_load_repositories
+from ...apps.load.adapters import queries as query_load_repositories
 from ...apps.load.adapters import units_of_work as load_unit_of_work
 
 
@@ -13,7 +14,7 @@ def inject_config(binder: inject.Binder):
     )
 
     binder.bind_to_constructor(
-        "save_data_repository", load_repositories.DjangoDataDomainRepository
+        "query_data_repository", query_load_repositories.DjangoDataQueryRepository
     )
 
 inject.configure(inject_config, once=True)
