+++ backend/modules/load/services/commands/commands.py
@@ -3,8 +3,8 @@ from ...domain.ports import units_of_work
 
 
 def save(
-    command: domain_commands.SaveData,
     unit_of_work: units_of_work.AbstractDataUnitOfWork,
+    command: domain_commands.SaveData,
 ):
     with unit_of_work:
         unit_of_work.data.create(command.output_data)
