+++ backend/modules/load/services/queries/ports/__init__.py
@@ -1,0 +1,3 @@
+from .repositories import AbstractDataQueryRepository
+
+__all__ = "AbstractDataQueryRepository"
